Refactor: 피드백 적용
1. 기능별로 파일 분리
2. 이벤트 핸들러 활용하여 this 사용 최소화
3. HTML 시멘틱 태그로 수정
4. form 활용
5. 뷰와 로직 분리
