+++ js/main.js
@@ -1,0 +1,11 @@
+import {
+  setStorage,
+  createItems,
+  setupFormHandlers,
+} from "./vendingMachine.js";
+
+document.addEventListener("DOMContentLoaded", function () {
+  setStorage();
+  createItems();
+  setupFormHandlers();
+});
